Changed corpus-main to use doubles and to print to 15 decimal places

+++ src/corpus/corpus-main.c
@@ -23,12 +23,12 @@ int main(void) {
     int i, j;
     for(i = 0; i < root->amount; i++) {
         corpus_chain * chain = root->root + i;
-        float total_seen = chain->seen_total;
+        unsigned long int total_seen = chain->seen_total;
         int value = chain->value;
         for(j = 0; j < chain->corpus_amount; j++) {
             corpus_node * node = chain->corpus + j;
-            float probability = (float)node->seen / total_seen;
-            printf("%c:%c:%f\n",value, node->other, probability);
+            double probability = (double)node->seen / (double)total_seen;
+            printf("%c:%c:%.15f\n",value, node->other, probability);
         }
     }
     free_chain(root);
